Update App.jsx
Changed signUP formular for name to be required and checked when submited, then added updateProfil functionality, names now display corretly , howewer there is bug , Profile pic is not showing , if you try to register without uploading picture there will be error that it cannot read property of null  but you are still logged in and it works but pic wont show, but the main issue is solved, so maybe open another issue for this bug. Because the eror will be gone if you upload the picture but it still wont show.